refactor folder names for consistancy
