+++ src/component/other/DuckDuckGo/index.jsx
